move 'case list form' tests to own file
