+++ corehq/apps/app_manager/tests/__init__.py
@@ -26,6 +26,7 @@ try:
     from .test_broken_build import *
     from .test_translations import *
     from corehq.apps.app_manager.tests.test_case_meta import *
+    from corehq.apps.app_manager.tests.test_case_list_form import *
 except ImportError, e:
     # for some reason the test harness squashes these so log them here for clarity
     # otherwise debugging is a pain
