+++ corehq/apps/app_manager/tests/data/case_list_form/case-list-form-suite-no-media-partial.xml
