+++ corehq/apps/app_manager/tests/data/case_list_form/case-list-form-suite-parent-child-basic.xml
