+++ corehq/apps/app_manager/tests/data/case_list_form/case_list_form_basic_form.xml
